Moving tracking interfaces to a subdir, making them inherit from a base class, renaming so they don't conflict with system libraries (like the soon-to-be-used python-fedex's fedex.py).
